Separate analyzed job suspension test suite
